+++ src/agentscope/model/_orcarouter/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""The OrcaRouter chat model modules."""
+
+from ._model import OrcaRouterChatModel
+
+__all__ = [
+    "OrcaRouterChatModel",
+]